requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.1
moviepy==1.0.3
pillow==9.5.0
numpy==1.26.4
requests==2.31.0
gTTS==2.5.1
srt==3.5.3
